Solution to Find Divisibility by 3 or 5

+++ Find Divisibility by 3 or 5@@6788dd63490637cfc44e1ba0/code.c
@@ -1,7 +1,7 @@
 #include <stdio.h>
 
 int main() {
-    int c;
+    float c;
     scanf("%d", &c);
     if ((c%3==0&&c%5==0)) {printf("Divisible by Both");} 
     else if(c%3==0) { printf("Divisible by 3");}
